requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.3.0
opencensus==0.7.10
opencensus-ext-azure==1.0.4
opencensus-ext-logging==0.1.0
requests==2.24.0
